refactor: split messaging_cli_flags tests for size compliance
