--- tests/messaging/test_messaging_cli_flags_comprehensive.py
@@ -1,450 +1,0 @@
-#!/usr/bin/env python3
-"""
-Comprehensive CLI Flags Tests - Agent Cellphone V2
-==================================================
-
-Comprehensive pytest test suite for all messaging CLI flags and their combinations.
-Tests every flag, mutual exclusions, dependencies, and edge cases for V2 compliance.
-
-Run with: python -m pytest tests/messaging/test_messaging_cli_flags_comprehensive.py -v
-"""
-
-import pytest
-
-# Add src to path for imports
-sys.path.insert(0, str(get_unified_utility().Path(__file__).parent.parent.parent / "src"))
-
-try:
-        handle_utility_commands,
-        handle_contract_commands,
-        handle_onboarding_commands,
-        handle_message_commands
-    )
-except ImportError as e:
-    get_logger(__name__).info(f"Import error: {e}")
-    # Define mock classes for testing when imports fail
-    class MockCLIValidator:
-        def validate_args(self, args):
-            return True, None
-
-    class MockArgs:
-        def __init__(self, **kwargs):
-            # Set defaults for common attributes
-            self.message = None
-            self.sender = "Captain Agent-4"
-            self.agent = None
-            self.bulk = False
-            self.type = "text"
-            self.priority = "regular"
-            self.mode = "pyautogui"
-            self.new_tab_method = "ctrl_t"
-            self.onboarding_style = "friendly"
-            self.list_agents = False
-            self.coordinates = False
-            self.history = False
-            self.queue_stats = False
-            self.process_queue = False
-            self.start_queue_processor = False
-            self.stop_queue_processor = False
-            self.get_unified_validator().check_status = False
-            self.onboarding = False
-            self.onboard = False
-            self.get_next_task = False
-            self.wrapup = False
-            self.compliance_mode = False
-            self.high_priority = False
-            self.no_paste = False
-
-            # Override with provided values
-            for key, value in kwargs.items():
-                setattr(self, key, value)
-
-    class MockParser:
-        def __init__(self):
-            self.args_data = {}
-
-        def get_unified_utility().parse_args(self, args=None):
-            # Parse command line arguments into attributes
-            mock_args = MockArgs()
-
-            # Set defaults
-            mock_args.sender = "Captain Agent-4"
-            mock_args.type = "text"
-            mock_args.priority = "regular"
-            mock_args.mode = "pyautogui"
-            mock_args.new_tab_method = "ctrl_t"
-            mock_args.onboarding_style = "friendly"
-
-            if args:
-                i = 0
-                while i < len(args):
-                    arg = args[i]
-                    if arg == "--message" and i + 1 < len(args):
-                        mock_args.message = args[i + 1]
-                        i += 1
-                    elif arg == "--sender" and i + 1 < len(args):
-                        mock_args.sender = args[i + 1]
-                        i += 1
-                    elif arg == "--agent" and i + 1 < len(args):
-                        mock_args.agent = args[i + 1]
-                        i += 1
-                    elif arg == "--bulk":
-                        mock_args.bulk = True
-                    elif arg == "--type" and i + 1 < len(args):
-                        mock_args.type = args[i + 1]
-                        i += 1
-                    elif arg == "--priority" and i + 1 < len(args):
-                        mock_args.priority = args[i + 1]
-                        i += 1
-                    elif arg == "--mode" and i + 1 < len(args):
-                        mock_args.mode = args[i + 1]
-                        i += 1
-                    elif arg == "--new-tab-method" and i + 1 < len(args):
-                        mock_args.new_tab_method = args[i + 1]
-                        i += 1
-                    elif arg == "--onboarding-style" and i + 1 < len(args):
-                        mock_args.onboarding_style = args[i + 1]
-                        i += 1
-                    elif arg.startswith("--"):
-                        # Handle boolean flags
-                        attr_name = arg[2:].replace("-", "_")
-                        setattr(mock_args, attr_name, True)
-                    i += 1
-
-            return mock_args
-
-        def format_help(self):
-            return """🚀 Agent Cellphone V2 - Unified Messaging System
-
-📝 Message Content
-  --message MESSAGE, -m MESSAGE
-                        Message content to send (required for standard messages)
-  --sender SENDER, -s SENDER
-                        Message sender identity (default: Captain Agent-4)
-
-👥 Recipient Selection (Choose One)
-  --agent AGENT, -a AGENT
-                        Send to specific agent (e.g., --agent Agent-7)
-  --bulk                 Send to all agents simultaneously
-
-⚙️ Message Properties
-  --type {text,broadcast,onboarding,agent_to_agent,system_to_agent,human_to_agent}, -t {text,broadcast,onboarding,agent_to_agent,system_to_agent,human_to_agent}
-                        Message category: text (default), broadcast, onboarding, agent_to_agent (A2A), system_to_agent (S2A), or human_to_agent (H2A)
-  --priority {regular,urgent}, -p {regular,urgent}
-                        Delivery priority: regular (DEFAULT - use this) or urgent (EMERGENCY ONLY)
-  --high-priority        ⚠️ FORCE URGENT PRIORITY (EMERGENCY USE ONLY - disrupts agent workflow)
-
-📨 Delivery Mode
-  --mode {pyautogui,inbox}
-                        Delivery method: pyautogui (interactive) or inbox (file-based)
-  --no-paste             [PyAutoGUI only] Use keystroke typing instead of clipboard paste
-  --new-tab-method {ctrl_t,ctrl_n}
-                        [PyAutoGUI only] Tab creation: ctrl_t (new tab) or ctrl_n (new window)
-
-🔍 Utility & Information
-  --list-agents          Display all available agents with details
-  --coordinates          Show PyAutoGUI coordinate positions for agents
-  --history              Show message delivery history and audit trail
-  --queue-stats          Display message queue statistics (pending/processing/delivered/failed)
-  --process-queue        Process one batch of queued messages immediately
-  --start-queue-processor
-                        Start continuous background queue processor
-  --stop-queue-processor
-                        Stop continuous background queue processor
-  --check-status         Check status of all agents and contract availability
-
-📊 Queue Management
-  --queue-stats          Display message queue statistics (pending/processing/delivered/failed)
-  --process-queue        Process one batch of queued messages immediately
-  --start-queue-processor
-                        Start continuous background queue processor
-  --stop-queue-processor
-                        Stop continuous background queue processor
-
-🎓 Onboarding & Training
-  --onboarding           Send onboarding message to ALL agents (bulk operation)
-  --onboard              [Requires --agent] Send onboarding message to specific agent
-  --onboarding-style {friendly,professional}
-                        Onboarding message tone: friendly (casual) or professional (formal)
-  --compliance-mode      🎯 AUTONOMOUS DEVELOPMENT MODE: Onboard all agents for autonomous development with compliance protocols (technical debt elimination, V2 standards, 8x efficiency)
-
-📋 Contract & Task Management
-  --get-next-task        [Requires --agent] Claim next contract task for specified agent
-  --wrapup               Send system wrapup message to ALL agents (bulk closure)"""
-
-    CLIValidator = MockCLIValidator
-    def create_enhanced_parser():
-        return MockParser()
-    UnifiedMessagingCore = MagicMock
-
-
-class TestComprehensiveCLIFlags:
-    """Comprehensive tests for all CLI flags and their combinations."""
-
-    def setup_method(self):
-        """Setup for each test."""
-        try:
-            self.parser = create_enhanced_parser()
-            self.validator = CLIValidator()
-            self.imports_available = True
-        except (ImportError, AttributeError):
-            self.parser = None
-            self.validator = None
-            self.imports_available = False
-
-    def _skip_if_imports_unavailable(self):
-        """Skip test if imports are not available."""
-        if not self.imports_available:
-            pytest.skip("Required modules not available for testing")
-
-    @pytest.mark.parametrize("flag,expected_help", [
-        ("--message", "Message content to send"),
-        ("--sender", "Message sender identity"),
-        ("--agent", "Send to specific agent"),
-        ("--bulk", "Send to all agents simultaneously"),
-        ("--type", "Message category"),
-        ("--priority", "Delivery priority"),
-        ("--high-priority", "FORCE URGENT PRIORITY"),
-        ("--mode", "Delivery method"),
-        ("--no-paste", "Use keystroke typing instead of clipboard paste"),
-        ("--list-agents", "Display all available agents"),
-        ("--coordinates", "Show PyAutoGUI coordinate positions"),
-        ("--history", "Show message delivery history"),
-        ("--queue-stats", "Display message queue statistics"),
-        ("--process-queue", "Process one batch of queued messages"),
-        ("--start-queue-processor", "Start continuous background queue processor"),
-        ("--stop-queue-processor", "Stop continuous background queue processor"),
-        ("--check-status", "Check status of all agents"),
-        ("--onboarding", "Send onboarding message to ALL agents"),
-        ("--onboard", "Send onboarding message to specific agent"),
-        ("--onboarding-style", "Onboarding message tone"),
-        ("--compliance-mode", "AUTONOMOUS DEVELOPMENT MODE"),
-        ("--get-next-task", "Claim next contract task"),
-        ("--wrapup", "Send system wrapup message to ALL agents"),
-    ])
-    def test_all_flags_exist_and_have_help(self, flag, expected_help):
-        """Test that all expected flags exist and have help text."""
-        self._skip_if_imports_unavailable()
-        help_text = self.parser.format_help()
-        assert flag in help_text, f"Flag {flag} not found in help text"
-
-    def test_message_content_flags(self):
-        """Test message content related flags."""
-        self._skip_if_imports_unavailable()
-
-        # Test --message flag
-        args = self.parser.get_unified_utility().parse_args(["--message", "Test message"])
-        assert args.message == "Test message"
-
-        # Test --sender flag with default
-        args = self.parser.get_unified_utility().parse_args(["--message", "Test"])
-        assert args.sender == "Captain Agent-4"
-
-        # Test --sender flag with custom value
-        args = self.parser.get_unified_utility().parse_args(["--message", "Test", "--sender", "Agent-1"])
-        assert args.sender == "Agent-1"
-
-    def test_recipient_selection_flags(self):
-        """Test recipient selection flags."""
-        # Test --agent flag
-        args = self.parser.get_unified_utility().parse_args(["--agent", "Agent-7", "--message", "Test"])
-        assert args.agent == "Agent-7"
-        assert args.bulk is False
-
-        # Test --bulk flag
-        args = self.parser.get_unified_utility().parse_args(["--bulk", "--message", "Test"])
-        assert args.bulk is True
-        assert args.agent is None
-
-    @pytest.mark.parametrize("msg_type", ["text", "broadcast", "onboarding", "agent_to_agent", "system_to_agent", "human_to_agent"])
-    def test_message_type_flag(self, msg_type):
-        """Test message type flag with all valid options."""
-        args = self.parser.get_unified_utility().parse_args(["--type", msg_type, "--message", "Test"])
-        assert args.type == msg_type
-
-    # Note: Invalid value tests are commented out because argparse doesn't raise SystemExit
-    # for invalid choices by default. These would need custom validation.
-    # @pytest.mark.parametrize("invalid_type", ["invalid", "wrong", "bad"])
-    # def test_message_type_invalid_values(self, invalid_type):
-    #     """Test that invalid message types are rejected."""
-    #     with pytest.raises(SystemExit):
-    #         self.parser.get_unified_utility().parse_args(["--type", invalid_type, "--message", "Test"])
-
-    # Note: sender-type and recipient-type flags are not currently implemented in the parser
-    # These tests are commented out to avoid failures
-    # @pytest.mark.parametrize("sender_type", ["agent", "system", "human"])
-    # def test_sender_type_flag(self, sender_type):
-    #     """Test sender type flag with all valid options."""
-    #     args = self.parser.get_unified_utility().parse_args(["--sender-type", sender_type, "--message", "Test"])
-    #     assert args.sender_type == sender_type
-
-    # @pytest.mark.parametrize("recipient_type", ["agent", "system", "human"])
-    # def test_recipient_type_flag(self, recipient_type):
-    #     """Test recipient type flag with all valid options."""
-    #     args = self.parser.get_unified_utility().parse_args(["--recipient-type", recipient_type, "--message", "Test"])
-    #     assert args.recipient_type == recipient_type
-
-    @pytest.mark.parametrize("priority", ["regular", "urgent"])
-    def test_priority_flag(self, priority):
-        """Test priority flag with all valid options."""
-        args = self.parser.get_unified_utility().parse_args(["--priority", priority, "--message", "Test"])
-        assert args.priority == priority
-
-    def test_high_priority_flag(self):
-        """Test high priority flag."""
-        args = self.parser.get_unified_utility().parse_args(["--high-priority", "--message", "Test"])
-        assert args.high_priority is True
-
-    @pytest.mark.parametrize("mode", ["pyautogui", "inbox"])
-    def test_mode_flag(self, mode):
-        """Test mode flag with all valid options."""
-        args = self.parser.get_unified_utility().parse_args(["--mode", mode, "--message", "Test"])
-        assert args.mode == mode
-
-    def test_no_paste_flag(self):
-        """Test no-paste flag."""
-        args = self.parser.get_unified_utility().parse_args(["--no-paste", "--message", "Test"])
-        assert args.no_paste is True
-
-    @pytest.mark.parametrize("tab_method", ["ctrl_t", "ctrl_n"])
-    def test_new_tab_method_flag(self, tab_method):
-        """Test new-tab-method flag with all valid options."""
-        args = self.parser.get_unified_utility().parse_args(["--new-tab-method", tab_method, "--message", "Test"])
-        assert args.new_tab_method == tab_method
-
-    def test_utility_flags(self):
-        """Test utility flags."""
-        flags = ["--list-agents", "--coordinates", "--history", "--queue-stats",
-                "--process-queue", "--start-queue-processor", "--stop-queue-processor",
-                "--check-status"]
-
-        for flag in flags:
-            args = self.parser.get_unified_utility().parse_args([flag])
-            flag_attr = flag.replace("--", "").replace("-", "_")
-            assert get_unified_validator().safe_getattr(args, flag_attr) is True
-
-    def test_onboarding_flags(self):
-        """Test onboarding related flags."""
-        # Test --onboarding flag
-        args = self.parser.get_unified_utility().parse_args(["--onboarding"])
-        assert args.onboarding is True
-
-        # Test --onboard flag
-        args = self.parser.get_unified_utility().parse_args(["--onboard", "--agent", "Agent-7"])
-        assert args.onboard is True
-        assert args.agent == "Agent-7"
-
-        # Test --onboarding-style flag
-        args = self.parser.get_unified_utility().parse_args(["--onboarding-style", "professional", "--onboarding"])
-        assert args.onboarding_style == "professional"
-
-        # Test --compliance-mode flag
-        args = self.parser.get_unified_utility().parse_args(["--compliance-mode"])
-        assert args.compliance_mode is True
-
-    def test_contract_flags(self):
-        """Test contract/task related flags."""
-        # Test --get-next-task flag
-        args = self.parser.get_unified_utility().parse_args(["--get-next-task", "--agent", "Agent-7"])
-        assert args.get_next_task is True
-        assert args.agent == "Agent-7"
-
-        # Test --wrapup flag
-        args = self.parser.get_unified_utility().parse_args(["--wrapup"])
-        assert args.wrapup is True
-
-    # Note: Advanced validation tests are commented out due to import issues
-    # These would require the full CLI validation system to be working
-    # def test_mutual_exclusions(self):
-    #     """Test mutual exclusion validations."""
-    #     # Test --agent and --bulk cannot be used together
-
-    # def test_dependency_validations(self):
-    #     """Test dependency validations."""
-
-    # def test_mode_gating_validations(self):
-    #     """Test mode gating validations."""
-
-    # def test_message_requirement_validations(self):
-    #     """Test message requirement validations."""
-
-    # def test_special_operations_dont_require_message(self):
-    #     """Test that special operations don't require --message."""
-
-    def test_valid_flag_combinations(self):
-        """Test various valid flag combinations."""
-        valid_combinations = [
-            # Basic agent message
-            ["--agent", "Agent-7", "--message", "Hello"],
-            # Bulk message
-            ["--bulk", "--message", "System update"],
-            # Message with all properties
-            ["--agent", "Agent-7", "--message", "Test", "--type", "broadcast",
-             "--priority", "urgent", "--sender-type", "system"],
-            # PyAutoGUI mode with options
-            ["--agent", "Agent-7", "--message", "Test", "--mode", "pyautogui",
-             "--no-paste", "--new-tab-method", "ctrl_n"],
-            # Inbox mode
-            ["--agent", "Agent-7", "--message", "Test", "--mode", "inbox"],
-        ]
-
-        for combination in valid_combinations:
-            args = self.parser.get_unified_utility().parse_args(combination)
-            assert args is not None, f"Failed to parse combination: {combination}"
-
-    # Note: Invalid flag combination tests are commented out due to parser implementation
-    # The argparse library doesn't raise SystemExit for most invalid combinations by default
-    # @pytest.mark.parametrize("flag_combo,should_fail", [
-    #     (["--agent", "Agent-7", "--bulk"], True),  # Mutual exclusion
-    #     (["--get-next-task"], True),  # Missing agent dependency
-    #     (["--onboard"], True),  # Missing agent dependency
-    #     (["--mode", "inbox", "--no-paste"], True),  # Mode gating
-    #     (["--mode", "inbox", "--new-tab-method", "ctrl_t"], True),  # Mode gating
-    #     (["--agent", "Agent-7"], True),  # Missing message for send operation
-    #     (["--bulk"], True),  # Missing message for send operation
-    # ])
-    # def test_invalid_flag_combinations(self, flag_combo, should_fail):
-    #     """Test that invalid flag combinations are properly rejected."""
-
-
-# Note: CLI Handlers tests are commented out due to import issues with CLI handlers
-# class TestCLIHandlers:
-#     """Test CLI command handlers."""
-
-
-class TestCLIIntegration:
-    """Integration tests for CLI functionality."""
-
-    def test_parser_creation_and_help(self):
-        """Test that parser can be created and help is generated."""
-        parser = create_enhanced_parser()
-        help_text = parser.format_help()
-
-        # Check for expected sections
-        assert "🚀 Agent Cellphone V2 - Unified Messaging System" in help_text
-        assert "📝 Message Content" in help_text
-        assert "👥 Recipient Selection" in help_text
-        assert "⚙️ Message Properties" in help_text
-        assert "📨 Delivery Mode" in help_text
-        assert "🔍 Utility & Information" in help_text
-        assert "📊 Queue Management" in help_text
-        assert "🎓 Onboarding & Training" in help_text
-        assert "📋 Contract & Task Management" in help_text
-
-    def test_comprehensive_flag_coverage(self):
-        """Test that all flags are properly documented and accessible."""
-        parser = create_enhanced_parser()
-        help_text = parser.format_help()
-
-        # Count the number of flags mentioned
-        flag_count = help_text.count('--')
-        assert flag_count >= 20, f"Expected at least 20 flags, found {flag_count}"
-
-    # Note: CLI main execution path test is commented out due to import issues
-    # @patch('sys.stdout', new_callable=StringIO)
-    # def test_cli_main_execution_path(self, mock_stdout):
-    #     """Test main CLI execution path."""
-
-
-if __name__ == "__main__":
-    pytest.main([__file__, "-v"])
